entrans/requirements.txt
Flask==3.0.3
Werkzeug==3.0.6

# Dashboard and visualization
dash==3.0.4
dash-bootstrap-components==2.0.2
plotly==6.0.1

# Data processing
pandas==2.2.3
numpy==2.2.5
pyarrow==20.0.0

# Scientific computing
scipy==1.15.3

# HTTP requests
requests==2.32.3

# Required by your modules
beautifulsoup4==4.13.4
lxml==5.4.0
Jinja2==3.1.6
click==8.2.0
itsdangerous==2.2.0
MarkupSafe==3.0.2

# Dashboard additional dependencies
matplotlib==3.10.3
kiwisolver==1.4.8
pillow==11.2.1
fonttools==4.58.4

# PySAM for solar simulation
NREL-PySAM==7.0.0
google-cloud-storage==2.14.0

joblib>=1.3.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0

python-dateutil>=2.8.0
gunicorn>=21.2.0